% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullRequests.R
\name{getEVARATEs}
\alias{getEVARATEs}
\title{Gets all EVARATE survey responses from MediSkill}
\usage{
getEVARATEs(mediConnection)
}
\arguments{
\item{mediConnection}{The API connection}
}
\value{
A data.frame with the results
}
\description{
Gets all EVARATE survey responses from MediSkill
}
