% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullRequests.R
\name{getOPERATEs}
\alias{getOPERATEs}
\title{Gets all OPERATE survey responses from MediSkill}
\usage{
getOPERATEs(mediConnection)
}
\arguments{
\item{mediConnection}{The API connection}
}
\value{
A data.frame with the results
}
\description{
Gets all OPERATE survey responses from MediSkill
}
