% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullRequests.R
\name{getProcedures}
\alias{getProcedures}
\title{Gets all logged procedures from MediSkill}
\usage{
getProcedures(mediConnection)
}
\arguments{
\item{mediConnection}{The API connection}
}
\value{
A data.frame with the results
}
\description{
Gets all logged procedures from MediSkill
}
