% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullRequests.R
\name{getSupervisors}
\alias{getSupervisors}
\title{Gets all supervisor details from MediSkill}
\usage{
getSupervisors(mediConnection)
}
\arguments{
\item{mediConnection}{The API connection}
}
\value{
A data.frame with the results
}
\description{
Gets all supervisor details from MediSkill
}
