% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullRequests.R
\name{getUsers}
\alias{getUsers}
\title{Gets all user details from MediSkill}
\usage{
getUsers(mediConnection)
}
\arguments{
\item{mediConnection}{The API connection}
}
\value{
A data.frame with the results
}
\description{
Gets all user details from MediSkill
}
