% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullRequests.R
\name{getVOTs}
\alias{getVOTs}
\title{Gets all VOT survey responses from MediSkill}
\usage{
getVOTs(mediConnection)
}
\arguments{
\item{mediConnection}{The API connection}
}
\value{
A data.frame with the results
}
\description{
Gets all VOT survey responses from MediSkill
}
