% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{mediskillConnection}
\alias{mediskillConnection}
\title{Creates a connection object for the MediSkill API}
\usage{
mediskillConnection(token, url = "https://mediskill.dk/api/")
}
\arguments{
\item{token}{The API token}

\item{url}{Base url for the api}
}
\value{
A connection object
}
\description{
Usage:
mediConn = mediskillConnection('your token here')
}
\details{
allProcedures = getProcedures(mediConn)
}
